pyproject.toml
[tool.poetry]
name = "whatsapp-medbot"
version = "0.1.0"
description = "A WhatsApp medical bot using LangChain and FastAPI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.3"
structlog = "^24.1.0"
alembic = "^1.13.1"
python-jose = "^3.3.0"
boto3 = "^1.34.34"
openai = "^1.12.0"
wikipedia = "^1.4.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
pyngrok = "^7.0.5"
langchain = "^0.3.25"  # Updated to be compatible with langchain-community
langchain-core = "^0.3"
langchain-community = "^0.3.23"
langchain-openai = "^0.3.16"
sqlalchemy = "^2.0.27"
pdfrw = "^0.4"
reportlab = "^3.6.12"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"